+++ fsl_mrs/tests/test_utils_quantify.py
@@ -221,12 +221,11 @@ def test_quantifyWater():
     Fitargs = {'ppmlim': [0.2, 5.2],
                'method': 'MH',
                'baseline_order': 0,
-               'metab_groups': [0],
-               'storeOldWrongMolality': True}
+               'metab_groups': [0]}
 
     res = fit_FSLModel(mrs, **Fitargs)
 
-    tissueFractions = {'GM': 0.6, 'WM': 0.4, 'CSF': 0.0}
+    tissueFractions = {'GM': 0.5, 'WM': 0.4, 'CSF': 0.1}
     TE = 0.03
     TR = 20
     T2dict = {'H2O_GM': 0.110,
@@ -246,14 +245,21 @@ def test_quantifyWater():
     res.calculateConcScaling(mrs,
                              q_info,
                              internal_reference=['Cr'],
-                             verbose=True)
+                             verbose=True,
+                             wrong_molality=True)
 
     print(res.getConc(scaling='raw'))
     print(res.getConc(scaling='internal'))
     print(res.getConc(scaling='molality'))
     print(res.getConc(scaling='molarity'))
 
+    old_molal_scaling = 1 / (tissueFractions['GM'] * 0.78 +
+                             tissueFractions['WM'] * 0.65 +
+                             tissueFractions['CSF'] * 0.97)
+    new_molal_scaling = old_molal_scaling * (1 - tissueFractions['CSF']) / \
+        (1 - tissueFractions['CSF'] * 0.97 * old_molal_scaling)
+
     assert np.allclose(res.getConc(scaling='internal'), 1.0)
-    assert np.allclose(res.getConc(scaling='molarity'), 10.78, atol=3E-1)
-    assert np.allclose(res.getConc(scaling='molality'), 10.78 * 1 / (0.6 * 0.78 + 0.4 * 0.65), atol=3E-1)
-    assert np.allclose(res.getConc(scaling='old_WRONG_molality'), 10.78 * 1 / (0.6 * 0.78 + 0.4 * 0.65), atol=3E-1)
+    assert np.allclose(res.getConc(scaling='molarity'), 12.77, atol=3E-1)
+    assert np.allclose(res.getConc(scaling='molality'), 12.77 * new_molal_scaling, atol=3E-1)
+    assert np.allclose(res.getConc(scaling='old_WRONG_molality'), 12.77 * old_molal_scaling, atol=3E-1)