+++ CHANGELOG.rst
@@ -1,9 +1,14 @@
 This document contains the FSL-MRS release history in reverse chronological order.
 
+2.5.0 (tbc)
+------------------------------
+- Bug fix in molality calculation, using ``1 / (1 - self.f_CSF)`` instead of ``1 / (1 - self.f_CSF)`` correction.
+- Added option in `fsl_mrs` and `fsl_mrsi` to output the previously used (<=2.4.17) incorrect molality concentration as well.
+
 2.4.17 (Friday 24th July 2026)
 ------------------------------
 - Bumped fslpy dependency
-- Removed Python 3.10 support. 
+- Removed Python 3.10 support
 
 2.4.16 (Monday 29th June 2026)
 ------------------------------